Adds parser tests
